bots/telegram/requirements.txt
Pyrogram
TgCrypto
python-dotenv